Fixes to morphing card component and restructuring
