+++ components/usage/index.ts
@@ -10,10 +10,10 @@ export default {
     ["modern-animation"]: {
         GravityTextSwapUsage: require("./modern-animation/gravity-text-swap.usage").default,
         TextWriterUsage: require("./modern-animation/text-writer.usage").default,
+        MorphingCardUsage: require("./modern-animation/morphing-card.usage").default,
     },
     components: {
         GradientFlowUsage: require("./components/gradient-flow.usage").default,
         ShimmerEffectUsage: require("./components/shimmer-effect.usage").default,
-        MorphingCardUsage: require("./components/morphing-card.usage").default,
     },
 }
