+++ content/docs/index.ts
@@ -8,7 +8,7 @@ import TextWriterConfig from "@/content/docs/modern-animation/text-writer.conten
 import ScaleInConfig from "@/content/docs/animation/scale-in.content";
 import ShimmerEffectConfig from "@/content/docs/components/shimmer-effect.content";
 import FlipCardConfig from "@/content/docs/animation/flip-card.content";
-import MorphingCardConfig from "@/content/docs/components/morphing-card.content";
+import MorphingCardConfig from "@/content/docs/modern-animation/morphing-card.content";
 import { DocsConfigType } from "@/types/docs-config.type";
 
 const DocsContentConfig: DocsConfigType = {
@@ -23,11 +23,11 @@ const DocsContentConfig: DocsConfigType = {
   "modern-animation": {
     "gravity-text-swap": GravityTextSwapConfig,
     "text-writer": TextWriterConfig,
+    "morphing-card": MorphingCardConfig,
   },
   components: {
     "gradient-flow": GradientFlowConfig,
     "shimmer-effect": ShimmerEffectConfig,
-    "morphing-card": MorphingCardConfig,
   },
 };
